Added Adreno Idler
By Arter97

It is an idling algorithm,
an efficient workaround for msm-adreno-tz's overheads.

Main goal is to lower the power consumptions while maintaining high-performance.

Since msm-adreno-tz tends to *not* use the lowest frequency even on idle,
Adreno idler replaces msm-adreno-tz's algorithm when it comes to
calculating idle frequency(mostly by ondemand's method).

The higher frequencies are not touched with this algorithm, so high-demanding
games will (most likely) not suffer from worsened performance.